DASHBOARD: spec feature for future dashboard page
Created my first feature spec and it seems to be working as expected,
which is nice.  It creates a user, logs in as that user, and tries
to verify that it is on the dashboard page.

Next, need to build out dashboard.

Using this as guide: http://everydayrails.com/2012/07/31/rails-admin-panel-from-scratch.html