Mailer: replaced missing configs for gmail
Using gmail to send application emails in dev environment.  In
order to do this the username and password need to be specified
in config/secrets.yml and these were missing.

For some reason I had also moved the domain into secrets.yml, but
that doesn't appear to be necessary so I put it back in the
environment configs.