garden: fix bug with creating new gardens
In my commit "spec: views/plants/edit.html.haml w/ factory (bd1b9a6)"
I made the user a required for a garden.  However my test for creating
gardens was failing.  This was because the garden _form.html.haml was
naming the user field :user instead of :user_id and the user object
could not be created.

For a long time I thought it was an issue with the params I had
whitelisted and I was focused on the garden_controller but the issue
was in the form.  Bummer.