SPEC CONF: specify database_cleaner settings
When trying to fix a test that was hitting unique constraint errs
for user emails I thought perhaps user data was not being removed
from the database during tests.  I searched around and learned that
database_cleaner is a pretty standard tool for avoiding such problems.

This config did not make the test run because I think my test was just
written wrong. However this seems like a standard configuration and
seems like more of an improvment than an issue, so I'm going to let
it ride.

See:
http://www.virtuouscode.com/2012/08/31/configuring-database_cleaner-with-rails-rspec-capybara-and-selenium/
https://forum.upcase.com/t/rspec-support-vs-helpers/4986