SPEC USER: use ffaker to generate unique emails
I was getting duplicate emails in my tests when generating multiple
gardens.

Specifically, the problem happened because "FactoryGirl.create(:garden)"
creates a new user.  Then "valid_attributes" (which is a function,
remmeber) creates a user using "FactoryGirl.create(:garden)" again.
Since the email address was set in the FactoryGirl garden factory,
it would be reused.

I was uncertain whether I should simply generate unique emails, or
write my tests in such a way that the gardens could share the same
user.  It was mostly a question of performance.  I put the question
out to SO here: http://stackoverflow.com/questions/40340934 and
decided it's fine to create a handful of user objects in the db.

So it looks like ffaker works fine for my purposes.