BOOTSTRAP: initial config for bootstrap styling
Started by following some of the instructions on
https://github.com/twbs/bootstrap-sass, but then switched over
to https://gorails.com/episodes/styling-with-bootstrap-sass

Pulling in bootstrap well.  I dropped the BS navbar template in
to a new partial and it is styled well.  The dropdown doesn't work
yet though and it looks like that's because application.js can't
load jquery.  Need to troubleshoot that.