SPEC USER: test for devise view
This was tricky because the default devise views provide the user
object as a variable named "resource" and I needed to replicate that.

I ended up stubbing out the resource variables as shown in this post:
http://stackoverflow.com/questions/14426746/testing-devise-views-with-rspec#19931625

I would have preferred to have taken the approach in this post:
http://stackoverflow.com/questions/14426746/testing-devise-views-with-rspec#31223305
I was able to put that module in the mixins folder and load it in the view
but I was still getting errors.  I'd still prefer that approach rather
than stubbing out the resources for every devise test, but not bothering
right now.