SPEC USER: fix test for user update action
The problem was that I was using

  `FactoryGirl::build(:user).attributes`

instead of

  `FactoryGirl::attributes_for(:user)`

The first one returns a hash of built database fields, whereas the
second one returns a hash of factory properties.  The database does
not have a `password` field (it encrypts the password) so the test
was not passing the required fields to the form. Thus Fail.