SPEC GARDENS_PLANTS: fixed controller tests
The controller tests were failing because the "invalid" plants were
being successfully saved.  That's because the only validation on
the model was the garden_id, which I removed.

It doesn't make sense to successfully save a blank record so I made
the plant name a required field, which satisfied the controller tests.