GARDENS_PLANTS: add plant via ajax
The plants are being added, but I'm concerned I've messed up the
flash behavior.  At the moment it is broken and the ajax flash
message is being displayed on page reload.  I followed this post
http://stackoverflow.com/questions/23967390/rails-flash-notice-via-ajax

I need to go back to that post and dig a little deeper into the
re-posted flash message.  But I'm also wondering if this code limits
my flash message to :notice info?  I assume there is :info and :warning
types too but I'm not sure they would show up with this layout/application
```
  <div id="flash">
    <% if notice.present? %>
        <%= render partial: "shared/notice_banner" %>
    <% end %>
  </div>
```