GARDENS_PLANTS: also update number of plants
Now that I'm updating the plant cards when a new one is added to
the garden (which is an un-realistic use-case, but fun coding), I
also want to update the number in "You're growing 39 plants".

I discovered that HAML kind of sucks for inline markup.  This blog
post is a good recap of it's shortcomings & work-arounds:

http://chriseppstein.github.io/blog/2010/02/08/haml-sucks-for-content/