GARDENS_PLANTS: js to auto-hide bootstrap alerts
This is the first bit of JS I've added to a rails app, so there's a
little learning here.  I'm not going to try to dig into coffee script
right now - I haven't heard great things so far and there's enough
things that I'm already learning.

I still need to troubleshoot why the alerts don't hide on log-out,
and get them to display the right alert for the provided flash
key.