GARDENS_PLANTS: hide alerts after logout
Learned a bit about turbo links today.  Apparently Devise uses them
(or is it Rails..) for the logout, display login and a few other
actions.  I'd like to learn how/when I should be using them myself
but I haven't dug in that far yet.

Anyway, the reason the alerts weren't being hidden on logout is that
the page doesn't actually reload and I wrappped my jquery in an
"on ready" block.  Since the page was pjax'ed in my js never ran.
This is resolved by "on turbolinks:load" instead of "on ready".